Projects/FundIssues: extract EditBounty
* Move EditBounty to own component

  This allows us to stop saving `detailsOpen` to shared bounty state and
instead keep it as local component state.

  We may be able to clean up more than what I've done here, but this is a
good start.

* Stop passing `bountySettings` as prop, since it can be retrieved
  directly within child components.
